Wrap objects created by clients with Py Wrappers
